finish v1

+++ exercises/test.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 #include <vector>
 int main() {
-    std::vector<int> vec{1, 2, 3, 4, 5};
+    std::vector<double> vec{1, 2, 3, 4, 5};
     std::cout << sizeof(vec) << std::endl;
-    std::cout << vec.size() * sizeof(int) << std::endl;
+    std::cout << vec.size() * sizeof(double) << std::endl;
     return 0;
 }
